requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.0.0
cryptography>=42.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.15
